ipc: do cyclic id allocation with ipcmni_extend mode
For ipcmni_extend mode, the sequence number space is only 7 bits.  So the
chance of id reuse is relatively high compared with the non-extended mode.

To alleviate this id reuse problem, the id allocation will be done
cyclically to cycle through all the 24-bit id space before wrapping around
when in ipcmni_extend mode.  This may cause the use of more memory in term
of the number of xa_nodes allocated as well as potentially more cachelines
used as the xa_nodes may be spread more sparsely in this case.

There is probably a slight memory and performance cost in doing cyclic id
allocation.  For applications that really need more than 32k unique IPC
identifiers, this is a small price to pay to avoid the id reuse problem.

As a result, the chance of id reuse should be even smaller in the
ipcmni_extend mode.  For users who worry about id reuse, they can turn on
ipcmni_extend mode, even if they don't need more than 32k IPC identifiers.

Link: http://lkml.kernel.org/r/1551379645-819-4-git-send-email-longman@redhat.com
Signed-off-by: Waiman Long <longman@redhat.com>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Cc: Davidlohr Bueso <dbueso@suse.de>
Cc: "Eric W. Biederman" <ebiederm@xmission.com>
Cc: Jonathan Corbet <corbet@lwn.net>
Cc: Kees Cook <keescook@chromium.org>
Cc: "Luis R. Rodriguez" <mcgrof@kernel.org>
Cc: Manfred Spraul <manfred@colorfullife.com>
Cc: Matthew Wilcox <willy@infradead.org>
Cc: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>