kdb: det rid of broken attempt to print CCVERSION in kdb summary
If you drop into kdb and type "summary", it prints out a line that says
this:

  ccversion  CCVERSION

...and I don't mean that it actually prints out the version of the C
compiler.  It literally prints out the string "CCVERSION".

The version of the C Compiler is already printed at boot up and it doesn't
seem useful to replicate this in kdb.  Let's just delete it.  We can also
delete the bit of the Makefile that called the C compiler in an attempt to
pass this into kdb.  This will remove one extra call to the C compiler at
Makefile parse time and (very slightly) speed up builds.

Link: http://lkml.kernel.org/r/20190323015227.245455-1-dianders@chromium.org
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Cc: Masahiro Yamada <yamada.masahiro@socionext.com>
Cc: Douglas Anderson <dianders@chromium.org>
Cc: Jason Wessel <jason.wessel@windriver.com>
Cc: Christophe Leroy <christophe.leroy@c-s.fr>
Cc: "Peter Zijlstra (Intel)" <peterz@infradead.org>
Cc: Kees Cook <keescook@chromium.org>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: "Gustavo A. R. Silva" <gustavo@embeddedor.com>
Cc: Nicholas Mc Guire <hofrat@osadl.org>,
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>